+++ sdk/src/org.graalvm.nativeimage/src/org/graalvm/nativeimage/package-info.java
@@ -40,11 +40,11 @@
  */
 /*
  @ApiInfo(
- group="Graal SDK"
+ group="GraalVM SDK"
  )
  */
 /**
- * The Graal-SDK native-image API allows to customize the native image generation, i.e., the
+ * The GraalVM SDK native-image API allows to customize the native image generation, i.e., the
  * ahead-of-time compilation of Java code to standalone executables. It also provides interfaces and
  * support classes that only work in the context of a native image.
  *