+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/io/package-info.java
@@ -40,11 +40,11 @@
  */
 /*
  @ApiInfo(
- group="Graal SDK"
+ group="GraalVM SDK"
  )
  */
 /**
- * Input/Output support for the Graal-SDK polyglot API.
+ * Input/Output support for the GraalVM SDK polyglot API.
  * <p>
  * The {@link org.graalvm.polyglot.io.FileSystem} allows an embedder to virtualize Input/Output
  * operations done by the {@code TruffleLanguage}s. A custom